all work so far. rpa reads work. APK somewhat
