+++ .gitignore
@@ -1,2 +1,2 @@
 DDLC-OSTPlayer-*.zip
-game/mod_assets/music_player/riffic-bold.ttf
+mod_assets/music_player/riffic-bold.ttf
