+++ mod_assets/music_player/NewUIHover.png
