+++ mod_assets/music_player/volumeHover.png
