+++ python-packages/ost_backend.py
@@ -1,0 +1,6 @@
+
+import renpy
+import ost_loader
+
+def file(fn):
+    return ost_loader.load(fn)
